--- src/components/DataTable/CustomCells/index.ts
@@ -1,2 +1,0 @@
-export * from './useDataTableCustomCells';
-export * from './dataTableCustomCells.types';
