+++ src/hooks/index.ts
@@ -1,6 +1,3 @@
-export * from './useTablePagination';
-export * from './useTableSelection';
-export * from './useTableSorting';
 export * from './mergeHandlers';
 export * from './onArrowIncrementation';
 export * from './onArrowNavigation';
